p1-ftds030-rmt-m2-wijayaa65/deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy
scikit-learn==1.4.1.post1
Pillow
plotly
feature-engine==1.7.0